[SimipleTodo] Add priorities to Todo items
Summary:
Allow setting of priorities in TodoItems.

Use a custom adapter to display the TodoItems in the list view, so that each
item in the list shows the item's task, priority, and the created time.

Test Plan:
Create Todo items with all three priorities: low, medium, and high. Edit the
items to make sure the priority in the spinner is selected as expected.